+++ gatsby-simplefolio/src/components/Contact.jsx
@@ -1,8 +1,8 @@
 import React, { useContext } from 'react';
 import Fade from 'react-reveal/Fade';
 import { Container } from 'react-bootstrap';
-import PortfolioContext from '../../context/context';
-import Title from '../Title/Title';
+import PortfolioContext from '../context/context';
+import Title from './Title';
 
 const Contact = () => {
   const { contact } = useContext(PortfolioContext);
