+++ gatsby-simplefolio/src/components/Footer.jsx
@@ -1,10 +1,10 @@
 import React, { useContext } from 'react';
 import { Container } from 'react-bootstrap';
 import { Link } from 'react-scroll';
-import PortfolioContext from '../../context/context';
-import GithubButtons from '../GithubButtons/GithubButtons';
+import PortfolioContext from '../context/context';
+import GithubButtons from './GithubButtons';
 
-import { githubButtons } from '../../mock/data';
+import { githubButtons } from '../mock/data';
 
 const Footer = () => {
   const { footer } = useContext(PortfolioContext);
