+++ gatsby-simplefolio/src/components/Projects.jsx
@@ -1,0 +1,123 @@
+import React, { useContext, useEffect, useState } from 'react';
+import Fade from 'react-reveal/Fade';
+import Tilt from 'react-tilt';
+import { Container, Row, Col } from 'react-bootstrap';
+import PortfolioContext from '../context/context';
+import Title from './Title';
+import ProjectImg from './Image/ProjectImg';
+
+const Projects = () => {
+  const { projects } = useContext(PortfolioContext);
+
+  const [isDesktop, setIsDesktop] = useState(false);
+  const [isMobile, setIsMobile] = useState(false);
+
+  useEffect(() => {
+    if (window.innerWidth > 769) {
+      setIsDesktop(true);
+      setIsMobile(false);
+    } else {
+      setIsMobile(true);
+      setIsDesktop(false);
+    }
+  }, []);
+
+  return (
+    <section id="where">
+      <Container>
+        <div className="project-wrapper">
+          <Title title="Where to see" />
+          {projects.map((project, i) => {
+            const { title, info, info2, url, repo, img, id } = project;
+            const projDesc = (
+              <Col lg={4} sm={12} key={`desc${id}`}>
+                <Fade
+                  left={isDesktop}
+                  bottom={isMobile}
+                  duration={1000}
+                  delay={250}
+                  distance="30px"
+                >
+                  <div className="project-wrapper__text">
+                    <h3 className="project-wrapper__text-title">{title || 'Project Title'}</h3>
+                    <div>
+                      <p>
+                        {info ||
+                          'Lorem ipsum dolor sit, amet consectetur adipisicing elit. Excepturi neque, ipsa animi maiores repellendu distinctioaperiam earum dolor voluptatum consequatur blanditiis inventore debitis fuga numquam voluptate architecto itaque molestiae.'}
+                      </p>
+                      <p className="mb-4">{info2 || ''}</p>
+                    </div>
+                    <a
+                      target="_blank"
+                      rel="noopener noreferrer"
+                      className="cta-btn cta-btn--hero"
+                      href={url || '#!'}
+                    >
+                      See Live
+                    </a>
+
+                    {repo && (
+                      <a
+                        target="_blank"
+                        rel="noopener noreferrer"
+                        className="cta-btn text-color-main"
+                        href={repo}
+                      >
+                        Source Code
+                      </a>
+                    )}
+                  </div>
+                </Fade>
+              </Col>
+            );
+            const projImg = (
+              <Col lg={8} sm={12} key={`img${id}`}>
+                <Fade
+                  right={isDesktop}
+                  bottom={isMobile}
+                  duration={1000}
+                  delay={500}
+                  distance="30px"
+                >
+                  <div className="project-wrapper__image">
+                    <a
+                      href={url || '#!'}
+                      target="_blank"
+                      aria-label="Project Link"
+                      rel="noopener noreferrer"
+                    >
+                      <Tilt
+                        options={{
+                          reverse: false,
+                          max: 8,
+                          perspective: 1000,
+                          scale: 1,
+                          speed: 300,
+                          transition: true,
+                          axis: null,
+                          reset: true,
+                          easing: 'cubic-bezier(.03,.98,.52,.99)',
+                        }}
+                      >
+                        <div data-tilt className="thumbnail rounded">
+                          <ProjectImg alt={title} filename={img} />
+                        </div>
+                      </Tilt>
+                    </a>
+                  </div>
+                </Fade>
+              </Col>
+            );
+            return (
+              <Row key={id}>
+                {isDesktop && i % 2 === 0 ? [projDesc, projImg] : [projImg, projDesc]}
+              </Row>
+            );
+          })}
+        </div>
+      </Container>
+    </section>
+  );
+};
+
+export default Projects;
