+++ gatsby-simplefolio/src/components/Title.jsx
