+++ gatsby-simplefolio/src/style/abstracts/_variables.scss
@@ -1,9 +1,12 @@
+//https://coolors.co/264653-2a9d8f-e9c46a-f4a261-e76f51
+// https://coolors.co/11151c-212d40-364156-7d4e57-d66853
 // COLORS
-$primary-color: #02aab0;
-$secondary-color: #00cdac;
-
+// $primary-color: #02aab0;
+// $secondary-color: #00cdac;
 // $primary-color: #7f449d;
 // $secondary-color: #793fc1;
+$primary-color: #e76f51;
+$secondary-color: #f4a261;
 
 $white-color: #fff;
