+++ gatsby-simplefolio/src/style/components/_buttons.scss
@@ -26,6 +26,7 @@
   background-image: linear-gradient(135deg, $primary-color 0%, $secondary-color 100%);
   -webkit-background-clip: text;
   -webkit-text-fill-color: transparent;
+  background-clip: text;
   border-width: 2px;
   border-style: solid;
   border-image: linear-gradient(135deg, $primary-color 0%, $secondary-color 100%);
