+++ gatsby-simplefolio/src/style/sections/_about.scss
@@ -1,10 +1,6 @@
-#about {
+#who {
   background-color: $primary-color;
-  background-image: linear-gradient(
-    135deg,
-    $primary-color 0%,
-    $secondary-color 100%
-  );
+  background-image: linear-gradient(135deg, $primary-color 0%, $secondary-color 100%);
   color: $white-color;
   height: 100%;
   border-top: 0px;
