+++ gatsby-simplefolio/src/style/sections/_projects.scss
@@ -1,4 +1,4 @@
-#projects {
+#where {
   background-color: $white-color;
   color: $dark-blue-text;
   margin-top: -15rem;
